+++ bin/Debug/net5.0/C_Lab1.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v5.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v5.0": {
+      "C_Lab1/1.0.0": {
+        "runtime": {
+          "C_Lab1.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "C_Lab1/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
